Add commands

+++ cogs/counting.py
@@ -6,6 +6,7 @@ class Counting(commands.Cog):
     def __init__(self, bot: commands.Bot) -> None:
         self.bot = bot
         self.current = None
+        self.active = True
         self.bot.loop.create_task(self.init())
 
     async def init(self) -> None:
@@ -20,7 +21,7 @@ class Counting(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message: Message) -> None:
-        if message.channel != self.channel or message.author.bot:
+        if not self.active or message.channel != self.channel or message.author.bot:
             return
         if (
             not message.content.isdigit()
@@ -33,6 +34,8 @@ class Counting(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message_edit(self, before: Message, after: Message) -> None:
+        if not self.active:
+            return
         if before == self.last:
             if (not after.content.isdigit()) or int(after.content) != self.current:
                 await after.delete()
@@ -41,10 +44,39 @@ class Counting(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message_delete(self, message: Message) -> None:
-        if message == self.last:
+        if self.active and message == self.last:
             self.last = await message.channel.send(
                 f"{self.current} (message from {message.author.name} was deleted)"
             )
+
+    @commands.group(invoke_without_command=True)
+    async def counting(self, ctx: commands.Context) -> None:
+        """
+        Commands related to the counting system
+        """
+        await ctx.send_help("counting")
+
+    @commands.has_permissions(manage_channels=True)
+    @counting.command(name="toggle")
+    async def _counting_toggle(self, ctx: commands.Context) -> None:
+        """
+        Toggle counting state
+        """
+        if self.active:
+            self.active = False
+            await ctx.send("Counting has been turned off.")
+        else:
+            self.active = True
+            await ctx.send("Counting has been turned on.")
+
+    @commands.has_permissions(manage_channels=True)
+    @counting.command(name="set")
+    async def _counting_set(self, ctx: commands.Context, number: int) -> None:
+        """
+        Set the current number
+        """
+        self.current = number
+        await ctx.send(f"Set current number to {number}")
 
 
 def setup(bot: commands.Bot) -> None:
